[MOD] default data manager security refactoring
